Start the API of DogWalkings

+++ db/migrate/20181129004345_create_dog_walkings.rb
@@ -1,12 +1,12 @@
 class CreateDogWalkings < ActiveRecord::Migration[5.2]
   def change
     create_table :dog_walkings do |t|
-      t.integer :status
+      t.integer  :status
       t.datetime :schedule_date
-      t.decimal :price
-      t.decimal :duration
-      t.float :latitude
-      t.decimal :longitude
+      t.decimal  :price
+      t.integer  :duration
+      t.decimal  :latitude
+      t.decimal  :longitude
       t.datetime :start_at
       t.datetime :end_at
